Moved the UI code out of the extension, into desktop-ui module.
